fix(user-provider): Do not access rehash function in L10 (#101)
This function was added in Laravel 11.
Currently, this causes an exception when using Fortify with 2FA and Laravel 10 together with WebAuthn.

In Fortify the function is called when it detects that it exists, for compatibility with Laravel 10. But since we implement this function in WebAuthn, we need to also add a check if this actually exists on the parent class.

So in Laravel 10 we will just ignore the function and do nothing. This should be a sufficient workaround to support both.